add prescaler to tx lrclk generation
Make the length of the lrclk cycle depend on the
'prescaler' input instead of the bit length.
This is to be able to adjust the samplerate, which
is not dependent on the bit length.
I.e. when the lrclk half-cycle is longer than the bit length,
only the bit length is still used, but the sample rate is
determined by the frequency of lrclk.